Added Missing Functionality
Added in functionality for the stepper motor, Error Message, Humidity Message, state transition reporting (with a helper function).
Also added in a separate stop button from the start button following an email from Bashira. Reenabled the interrupt for the start button and reprogrammed the not DISABLED states to respond to the stop button instead.
Code has not been tested with the circuit yet. Will commit again after testing with bugfixes + whitespace cleanup