Finished Code
Did troubleshooting with the circuit. Added a function for reporting the Temperature and Humidty, which also gets used during certain state transitions. added a serial print out for when the stepper motors get used (Also fixed the stepper motors so they work).  Fixed reportTransition to not report the time in hexadecimal. Cleaned up whitespace.
Should be the final version of the code.